resolved review, split apdapter
