+++ app/src/main/java/com/example/cocktaildb/data/model/IngredientItem.kt
@@ -1,0 +1,7 @@
+package com.example.cocktaildb.data.model
+
+
+data class IngredientItem(
+    var name: String,
+    var measure: String
+)
